Removed `MetaManager::ListTitles` because I couldn't get it to work. It's failing for an unknown reason, and there's insufficient resources to lead to the source issue. I'm also now questioning if the function would perform in the manner I was expecting it to.
`GameBrowser` is now being used to gather what I'm now thinking may be the actual game name.

Since the icon is being gathered from the same loaded data, I extracted that functionality from the submodule, so it could be reused for loading the name.

Has not yet been tested, so it's likely very broken at this commit.

Other notes:

* Added destructor to the new `Game` object to deallocate memory. This wasn't happening before, but to the best of my knowledge, I think it should (at least to be safe).

* Removed logging that wasn't being used

* Renamed `_groupList_` to `_groupItems_` so it won't get confused with `groupList` which is a totally different type