Improved how icons are being loaded
A struct now isn't ideal, so created a `Game` class to store and manage the data.

Icon memory alocation is now handled internal to that class through the setter.

I think the icon buffer size is being more accurately computed now too.

I think there was also a memory leak with the name data, so added `delete` to free it after getting the name from it.