More updates to further support the Borealis upgrade
Much of `GroupBrowser`'s code was removed because it's all either not needed anymore or it will be moved to `ModBrowser`.

`GroupBrowser` was also changed to a `TabFrame` since it originally was meant to operate similar to a tab frame. We'll see how well it works now. We probably won't need the custom `SubTabFrame` component with the additions/changes in the Borealis upgrade.